Use Vector<Byte> for fs.write_bytes
Align the stdlib, builtin typing, interpreter host bridge, and documentation around a byte-oriented write_bytes API.

This makes raw byte output match read_bytes and gives the boot wasm serializer plan a natural Vector<Byte> file boundary. The serializer plan is also updated to focus on the concrete path to success: reuse existing byte-building pieces, bring up direct wasm emission early, and remove the WAT bridge from the self-host critical path.

+++ src/types/env.rs
@@ -924,7 +924,7 @@ impl ValueEnv {
         env.builtins.insert(
             "__host_write_bytes".to_string(),
             MonoType::Function {
-                params: vec![MonoType::String, MonoType::Vector(Box::new(MonoType::Int))],
+                params: vec![MonoType::String, MonoType::Vector(Box::new(MonoType::Byte))],
                 ret: Box::new(MonoType::Void),
             },
         );
